requirements.txt
pynput
mss
requests
python-dotenv



